Revert "Move extension handling to new library"
This reverts commit 742c8cdf8d6cdfa076459146a648ba5835e69a54.
