+++ CMakeLists.txt
@@ -126,7 +126,6 @@ add_subdirectory(CesiumUtility)
 add_subdirectory(CesiumGltf)
 add_subdirectory(CesiumGeometry)
 add_subdirectory(CesiumGeospatial)
-add_subdirectory(CesiumJsonExtensions)
 add_subdirectory(CesiumJsonReader)
 add_subdirectory(CesiumGltfWriter)
 add_subdirectory(CesiumGltfReader)
